backend/pyproject.toml
[project]
name = "foresight-backend"
version = "0.1.0"
description = "Foresight - PR & Media Monitoring Platform"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Foresight Team" }
]

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "anthropic>=0.40.0",
    "psycopg2-binary>=2.9.0",
    "praw>=7.7.0",
    "tweepy>=4.14.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "crewai>=1.9.0",
    "neo4j>=5.20.0",
    "voyageai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
